Added some preliminary support for a single ethernet driver.

+++ u.h
@@ -3,7 +3,7 @@ typedef unsigned int u32;
 typedef unsigned short u16;
 typedef unsigned char u8;
 
-typedef long i64;
-typedef int i32;
-typedef short i16;
-typedef char i8;
+typedef long long s64;
+typedef int s32;
+typedef short s16;
+typedef char s8;
